// ProcMgt/ClientApp/src/app/models/item.ts
export class Item {

  itemID: string;
  itemTypeID: string; 
  itemDescription: string; 
  itemCode: string;
  currentQty: number; 
  initialQty: number;
  isActive: boolean;
  reOrderQuantity: number;
  userID: string;
  entryDateTime: string;

  itemTypeName: string;
  currenrtStockQty: string;

}